requirement.txt
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.13.1
anyio==4.3.0
apispec==6.5.0
argcomplete==3.2.3
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
Babel==2.14.0
backports.zoneinfo==0.2.1
bcrypt==4.1.2
blinker==1.7.0
cachelib==0.9.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
clickclick==20.10.2
colorama==0.4.6
colorlog==4.8.0
ConfigUpdater==3.2
connexion==2.14.2
cron-descriptor==1.4.3
croniter==2.0.2
cryptography==41.0.7
Deprecated==1.2.14
dill==0.3.1.1
dnspython==2.6.1
docutils==0.20.1
elasticsearch==6.3.1
email-validator==1.3.1
exceptiongroup==1.2.0
Flask==2.2.5
Flask-AppBuilder==4.3.11
Flask-Babel==2.0.0
Flask-Caching==2.1.0
Flask-Ext==0.1
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.5.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
flask-msearch==0.2.9.4
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WhooshAlchemy==0.56
Flask-WTF==1.2.1
frozenlist==1.4.1
fsspec==2024.2.0
google-re2==1.1
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.62.1
gunicorn==21.2.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.6
importlib-metadata==6.11.0
importlib-resources==5.13.0
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
lazy-object-proxy==1.10.0
limits==3.10.0
linkify-it-py==2.0.3
lockfile==0.12.2
Mako==1.3.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
marshmallow-oneofschema==3.1.1
marshmallow-sqlalchemy==0.26.1
mdit-py-plugins==0.4.0
mdurl==0.1.2
more-itertools==10.2.0
multidict==6.0.5
mysql-connector-python==8.3.0
numpy==1.24.4
opentelemetry-api==1.23.0
opentelemetry-exporter-otlp==1.23.0
opentelemetry-exporter-otlp-proto-common==1.23.0
opentelemetry-exporter-otlp-proto-grpc==1.23.0
opentelemetry-exporter-otlp-proto-http==1.23.0
opentelemetry-proto==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-semantic-conventions==0.44b0
ordered-set==4.1.0
packaging==23.2
pandas==2.0.3
pathspec==0.12.1
pendulum==3.0.0
pkgutil_resolve_name==1.3.10
pluggy==1.4.0
prison==0.2.1
protobuf==4.25.3
psutil==5.9.8
pycparser==2.21
Pygments==2.17.2
PyJWT==2.8.0
python-daemon==3.0.1
python-dateutil==2.9.0.post0
python-nvd3==0.15.0
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
requests==2.31.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rich==13.7.1
rich-argparse==1.4.0
rpds-py==0.18.0
setproctitle==1.3.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==1.4.52
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.1
sqlparse==0.4.4
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.4.0
text-unidecode==1.3
time-machine==2.14.0
typing_extensions==4.10.0
tzdata==2024.1
uc-micro-py==1.0.3
unicodecsv==0.14.1
universal-pathlib==0.1.4
urllib3==1.26.18
Werkzeug==2.2.3
Whoosh==2.7.4
wrapt==1.16.0
WTForms==3.1.2
yarl==1.9.4
zipp==3.17.0
